Revert to only defining one real repository, as lein deploy is unable to make proper use of it anyway:
lein deploy will always attempt to deploy to our mavenrepo/public, because
in settings.xml, it is defined as mirrorOf * (in maven, this doesn't apply
to distributionManagement).

lein pom does not generate distributionManagement, nor does it allow a
way to define packaging (=war).

Because of the above, we'll skip the lein pom/deploy step entirely, and rely
on using the mvn deploy-file command instead, which makes proper use of
our settings.xml.