Generate UUID for id
In both database and memory storage.

We hope that this will increase the stability of how messages are stored in
fleetdb. We suspect that the "count" "messages" query is failing some times.

Another theory is that our fleetdb connection handling is wrong, or not
suited for concurrency, but we'll explore this in the future.